requirements.txt
mcrcon==0.7.0
python-telegram-bot==20.7
PyYAML==6.0.1
